Default CodingWidget to WANG_2012 schema
This makes pretty much all of the hard problems go away.  When (or more likely if) support for multiple schemas is actually required, I think (having wrestled with a number of other approaches) that doing the whole thing with ajax and treating the entire JSON object as a single value from the ModelForm's point-of-view will be the way to go.